Refactored multi signin
