+++ tests/src/test/java/ca/bc/gov/open/jag/efiling/page/BasePage.java
@@ -20,4 +20,8 @@ public abstract class BasePage {
         PageFactory.initElements(this.driver, this);
     }
 
+    public void goTo(String url) {
+        this.driver.get(url);
+    }
+
 }
